# doc/openapi.yaml
openapi: 3.0.1
info:
  title: Time Object DB
  description: <b>Time Object DB API Contract</b> <br> Swagger Contract that describe
    the use of this database
  contact:
    name: Claudio Heidel Schemberger
    url: https://www.linkedin.com/in/mrcheidel/
  version: 1.0.9
servers:
- url: https://time-object-db.claudioheidel.repl.co/v1
...
paths:
...
  /databases:
    post:
      tags:
      - Databases
      summary: Service to create a new database.
      description: Service to create a new database.
      operationId: createDb
      responses:
        201:
          description: Created - The resource was created successfully and its identifier
            and link to the details are included in the body of the response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/databaseCreationResponse'
        400:
          description: Bad Request - If the resource is misspelled (badly constructed
            URI) / If the data in the request is not correct.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/400BadRequestError'
        500:
          description: Internal Server Error - Use it as a General System Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/500InternalServerError'
  /databases/{databaseId}:
    delete:
      tags:
      - Databases
      summary: Service to delete an complete database .
      description: Service to clear an complete database.
      operationId: deleteDb
      parameters:
      - name: token
        in: header
        description: Security Key
        required: true
        schema:
          type: string
      - name: databaseId
        in: path
        description: Database Identifier
        required: true
        schema:
          type: string
      responses:
        204:
          description: Created - The resource was created successfully and its identifier
            and link to the details are included in the body of the response.
          content: {}
        400:
          description: Bad Request - If the resource is misspelled (badly constructed
            URI) / If the data in the request is not correct.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/400BadRequestError'
        403:
          description: Forbidden - Invalid heade security token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/403ForbiddenError'
        404:
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/404NotFoundError'
        500:
          description: Internal Server Error - Use it as a General System Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/500InternalServerError'
...
components:
  schemas:
...
    databaseCreationResponse:
      type: object
      properties:
        databaseId:
          type: string
          description: Database Unique Identifier
          example: bcd9b4c0-1296-11ea-907c-11d417ac91a9
        key:
          type: string
          description: Key to be used as security token
          example: 570af3c05cc424283c0e6bf14ffda17b3a07bd7f5f2a0d172b436805f4dexxxx
      description: List of objects
...
    errorObject:
      title: errorObject
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: "400"
        description:
          type: string
          description: Error description
          example: Bad Request
        level:
          type: string
          description: Error level
          example: ERROR
          enum:
          - INFO
          - ERROR
          - WARNING
          - FATAL
        message:
          type: string
          description: Error message (detailed description)
          example: Invalid Request. Please, check the data in the request (QueryString
            Parameters and/or Body).
      description: Error details
      example:
        code: "400"
        description: Bad Request
        level: ERROR
        message: Invalid Request. Please, check the data in the request (QueryString
          Parameters and/or Body).
    400BadRequestError:
      title: 400BadRequestError
      type: object
      properties:
        errors:
          type: array
          description: Details of each of the errors
          items:
            $ref: '#/components/schemas/errorObject'
      description: List of errors
      example:
        errors:
        - code: "400"
          description: Bad Request
          level: ERROR
          message: Invalid Request. Please, check the data in the request (QueryString
            Parameters and/or Body).
    403ForbiddenError:
      title: 403ForbiddenError
      type: object
      properties:
        errors:
          type: array
          description: Details of each of the errors
          items:
            $ref: '#/components/schemas/errorObject'
      description: List of errors
      example:
        errors:
        - code: "403"
          description: Forbidden Error
          level: ERROR
          message: Forbidden Error. The user is not allowed to executed this operation.
    404NotFoundError:
      title: 404NotFoundError
      type: object
      properties:
        errors:
          type: array
          description: Details of each of the errors
          items:
            $ref: '#/components/schemas/errorObject'
      description: List of errors
      example:
        errors:
        - code: 404
          description: Resouce not found
          level: ERROR
          message: Resouce not found. Please, check the ResourceId specified.
    500InternalServerError:
      title: 500InternalServerError
      type: object
      properties:
        errors:
          type: array
          description: Details of each of the errors
          items:
            $ref: '#/components/schemas/errorObject'
      description: List of errors
      example:
        errors:
        - code: "500"
          description: Internal Server Error
          level: ERROR
          message: Internal Server Error. The server encountered an unexpected condition
            that prevented it from fulfilling the request.